requirements.txt
certifi==2020.6.20chardet==3.0.4furl==2.1.0idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'marshmallow==3.6.1orderedmultidict==1.0.1pyjwt==1.7.1requests==2.24.0six==1.15.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'